minor refactors to improve performance
