+++ store/app/src/main/java/br/com/stone/vianna/starstore/feature/itemList/ItemListActivity.kt
@@ -9,8 +9,8 @@ import android.widget.TextView
 import br.com.stone.vianna.starstore.entity.Item
 import br.com.stone.vianna.starstore.R
 import br.com.stone.vianna.starstore.baseClasses.BaseActivity
-import br.com.stone.vianna.starstore.helper.hide
-import br.com.stone.vianna.starstore.helper.show
+import br.com.stone.vianna.starstore.extensions.hide
+import br.com.stone.vianna.starstore.extensions.show
 import br.com.stone.vianna.starstore.feature.shoppingCart.ShoppingCartActivity
 import br.com.stone.vianna.starstore.feature.transactionHistory.TransactionHistoryActivity
 import kotlinx.android.synthetic.main.activity_item_list.*
@@ -22,7 +22,7 @@ class ItemListActivity : BaseActivity(), ItemListContract.View {
 
     private val presenter: ItemListContract.Presenter by inject { parametersOf(this) }
     private var textCartItemCount: TextView? = null
-    private var adapter = ItemsAdapter { item: Item ->
+    private val adapter = ItemsAdapter { item: Item ->
         presenter.onItemClicked(item)
     }
