+++ store/app/src/main/java/br/com/stone/vianna/starstore/feature/shoppingCart/ShoppingCartAdapter.kt
@@ -4,7 +4,7 @@ import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import br.com.stone.vianna.starstore.helper.toMoneyFormat
+import br.com.stone.vianna.starstore.extensions.toMoneyFormat
 import br.com.stone.vianna.starstore.R
 import br.com.stone.vianna.starstore.entity.Item
 import kotlinx.android.synthetic.main.view_item_cart.view.*
